imp: Change image to python buster

+++ entrypoint.sh
@@ -9,7 +9,11 @@ cp -r /github/main-branch/tests /github/workspace
 
 # Install deps and run pytest over the student source
 cd /github/workspace
-python3 -m pip install -r requirements.txt
+if [[ -f "dev-requirements.txt" ]]; then
+  python3 -m pip install -r dev-requirements.txt
+else
+  python3 -m pip install -r requirements.txt
+fi
 python3 -m pytest --json=/tmp/report.json
 
 # Run evaluator over pytest result assuring that the requirements file is the original
